Add base numeric property input
