+++ src/components/base-numeric-property-input/style.css
@@ -1,0 +1,3 @@
+:host {
+    display: inline-block;
+}
