+++ src/components/rotation-property-input/index.ts
@@ -1,53 +1,10 @@
-import { LitElement, css, unsafeCSS } from 'lit';
+import { css, unsafeCSS } from 'lit';
 import componentStyles from './style.css?inline';
 import { template } from './template';
-import { defineDraggableNumber } from '../draggable-number';
-import { definePropertyInput } from '../property-input';
+import { BaseNumericPropertyInput } from '../base-numeric-property-input';
 
-type DraggableNumberElement = HTMLElement & { value: number };
-
-export class RotationPropertyInput extends LitElement {
+export class RotationPropertyInput extends BaseNumericPropertyInput {
     static styles = css`${unsafeCSS(componentStyles)}`;
-
-    static properties = {
-        value: { type: Number, reflect: true },
-        min: { type: Number, reflect: true },
-        max: { type: Number, reflect: true },
-        disabled: { type: Boolean, reflect: true }
-    };
-
-    declare value: number;
-    declare min: number | null;
-    declare max: number | null;
-    declare disabled: boolean;
-
-    constructor() {
-        super();
-        if (!this.hasAttribute('value')) {
-            this.value = 0;
-        }
-        if (!this.hasAttribute('min')) {
-            this.min = null;
-        }
-        if (!this.hasAttribute('max')) {
-            this.max = null;
-        }
-        if (!this.hasAttribute('disabled')) {
-            this.disabled = false;
-        }
-    }
-
-    private _onNumberChange(e: Event) {
-        const val = (e.target as DraggableNumberElement).value;
-        this.value = val;
-        this.dispatchEvent(new Event('change'));
-    }
-
-    connectedCallback() {
-        super.connectedCallback();
-        definePropertyInput();
-        defineDraggableNumber();
-    }
 
     render() {
         return template(
